Use float as basis unit in shapes and add Rect2f and Rect2i

+++ src/test/java/org/terasology/math/geom/CircleTest.java
@@ -30,7 +30,7 @@ public class CircleTest {
     @Test
     public void testContains() {
         Circle c = new Circle(3, 5, 2);
-        assertTrue(c.contains(new Vector2d(2, 4)));
-        assertFalse(c.contains(new Vector2d(1, 4)));
+        assertTrue(c.contains(new Vector2f(2, 4)));
+        assertFalse(c.contains(new Vector2f(1, 4)));
     }
 }
